web/feed: Move /dependencies into /feed - backer view
