+++ clients/apps/web/src/pages/dependencies/[organization]/[repo]/index.tsx
@@ -1,50 +1,31 @@
-import Dashboard from '@/components/Dashboard'
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
-import type { NextLayoutComponentType } from 'next'
-import Head from 'next/head'
+import LoadingScreen from '@/components/Dashboard/LoadingScreen'
+import Layout from '@/components/Layout/EmptyLayout'
+import type { NextPageWithLayout } from '@/utils/next'
 import { useRouter } from 'next/router'
 import { ReactElement, useEffect } from 'react'
-import { useCurrentOrgAndRepoFromURL } from '../../../../hooks'
 
-const Page: NextLayoutComponentType = () => {
+const Page: NextPageWithLayout = () => {
   const router = useRouter()
-  const { organization: orgSlug, repo: repoSlug } = router.query
-  const key = `orgrepo-${orgSlug}-${repoSlug}` // use key to force reload of state
-  const { org, repo, isLoaded } = useCurrentOrgAndRepoFromURL()
-
   useEffect(() => {
-    if (isLoaded && !org && !repo) {
-      router.push('/dependencies')
-      return
-    }
-    if (isLoaded && org && !repo) {
-      router.push(`/dependencies/${org.name}`)
-      return
-    }
-  }, [isLoaded, org, repo, router])
-
-  if (!isLoaded) {
-    return <></>
-  }
+    router.push(`/feed`)
+  })
 
   return (
     <>
-      <Head>
-        <title>Polar{org && repo ? ` ${org.name}/${repo.name}` : ''}</title>
-      </Head>
-      <Dashboard
-        key={key}
-        org={org}
-        repo={repo}
-        isPersonal={false}
-        isDependencies={true}
-      />
+      <LoadingScreen>
+        <>Redirecting...</>
+      </LoadingScreen>
     </>
   )
 }
 
 Page.getLayout = (page: ReactElement) => {
-  return <Gatekeeper>{page}</Gatekeeper>
+  return (
+    <Gatekeeper>
+      <Layout>{page}</Layout>
+    </Gatekeeper>
+  )
 }
 
 export default Page
